perf: lazy load command action modules
This is intended to improve performance of commands by only loading the
modules required by the command being issued. Commands vary in their
dependencies, and some have very large dependency sets.

This commit doesn't cover all commands. There are still about 10 more to
update.
