+++ src/commands/api/api.mjs
@@ -10,7 +10,7 @@ import { chalk, error, exit, log, logJson } from '../../utils/command-helpers.mj
  * @param {import('commander').OptionValues} options
  * @param {import('../base-command.mjs').default} command
  */
-const apiCommand = async (apiMethod, options, command) => {
+export const apiCommand = async (apiMethod, options, command) => {
   const { api } = command.netlify
 
   if (options.list) {
@@ -48,23 +48,3 @@ const apiCommand = async (apiMethod, options, command) => {
     error(error_)
   }
 }
-
-/**
- * Creates the `netlify api` command
- * @param {import('../base-command.mjs').default} program
- * @returns
- */
-export const createApiCommand = (program) =>
-  program
-    .command('api')
-    .argument('[apiMethod]', 'Open API method to run')
-    .description(
-      `Run any Netlify API method
-For more information on available methods checkout https://open-api.netlify.com/ or run '${chalk.grey(
-        'netlify api --list',
-      )}'`,
-    )
-    .option('-d, --data <data>', 'Data to use')
-    .option('--list', 'List out available API methods', false)
-    .addExamples(['netlify api --list', `netlify api getSite --data '{ "site_id": "123456" }'`])
-    .action(apiCommand)
