+++ src/commands/init/init.mjs
@@ -1,5 +1,4 @@
 // @ts-check
-import { Option } from 'commander'
 import inquirer from 'inquirer'
 import isEmpty from 'lodash/isEmpty.js'
 
@@ -8,8 +7,8 @@ import getRepoData from '../../utils/get-repo-data.mjs'
 import { ensureNetlifyIgnore } from '../../utils/gitignore.mjs'
 import { configureRepo } from '../../utils/init/config.mjs'
 import { track } from '../../utils/telemetry/index.mjs'
-import { link } from '../link/index.mjs'
-import { sitesCreate } from '../sites/index.mjs'
+import { link } from '../link/link.mjs'
+import { sitesCreate } from '../sites/sites-create.mjs'
 
 const persistState = ({ siteInfo, state }) => {
   // Save to .netlify/state.json file
@@ -219,25 +218,3 @@ export const init = async (options, command) => {
 
   return siteInfo
 }
-
-/**
- * Creates the `netlify init` command
- * @param {import('../base-command.mjs').default} program
- * @returns
- */
-export const createInitCommand = (program) =>
-  program
-    .command('init')
-    .description(
-      'Configure continuous deployment for a new or existing site. To create a new site without continuous deployment, use `netlify sites:create`',
-    )
-    .option('-m, --manual', 'Manually configure a git remote for CI')
-    .option('--force', 'Reinitialize CI hooks if the linked site is already configured to use CI')
-    .addOption(
-      new Option(
-        '--gitRemoteName <name>',
-        'Old, prefer --git-remote-name. Name of Git remote to use. e.g. "origin"',
-      ).hideHelp(true),
-    )
-    .option('--git-remote-name <name>', 'Name of Git remote to use. e.g. "origin"')
-    .action(init)
