+++ src/commands/link/link.mjs
@@ -1,5 +1,4 @@
 // @ts-check
-import { Option } from 'commander'
 import inquirer from 'inquirer'
 import isEmpty from 'lodash/isEmpty.js'
 
@@ -330,24 +329,3 @@ export const link = async (options, command) => {
   }
   return siteData
 }
-
-/**
- * Creates the `netlify link` command
- * @param {import('../base-command.mjs').default} program
- * @returns
- */
-export const createLinkCommand = (program) =>
-  program
-    .command('link')
-    .description('Link a local repo or project folder to an existing site on Netlify')
-    .option('--id <id>', 'ID of site to link to')
-    .option('--name <name>', 'Name of site to link to')
-    .addOption(
-      new Option(
-        '--gitRemoteName <name>',
-        'Old, prefer --git-remote-name. Name of Git remote to use. e.g. "origin"',
-      ).hideHelp(true),
-    )
-    .option('--git-remote-name <name>', 'Name of Git remote to use. e.g. "origin"')
-    .addExamples(['netlify link', 'netlify link --id 123-123-123-123', 'netlify link --name my-site-name'])
-    .action(link)
