+++ src/commands/lm/lm-info.mjs
@@ -11,7 +11,7 @@ import {
 /**
  * The lm:info command
  */
-const lmInfo = async () => {
+export const lmInfo = async () => {
   const steps = [
     checkGitVersionStep,
     checkGitLFSVersionStep,
@@ -30,11 +30,3 @@ const lmInfo = async () => {
     // an error is already reported when a task fails
   }
 }
-
-/**
- * Creates the `netlify lm:info` command
- * @param {import('../base-command.mjs').default} program
- * @returns
- */
-export const createLmInfoCommand = (program) =>
-  program.command('lm:info', { hidden: true }).description('Show large media requirements information.').action(lmInfo)
