+++ src/commands/lm/lm.mjs
@@ -1,34 +1,68 @@
 // @ts-check
-import { createLmInfoCommand } from './lm-info.mjs'
-import { createLmInstallCommand } from './lm-install.mjs'
-import { createLmSetupCommand } from './lm-setup.mjs'
-import { createLmUninstallCommand } from './lm-uninstall.mjs'
-
 /**
- * The lm command
- * @param {import('commander').OptionValues} options
- * @param {import('../base-command.mjs').default} command
- */
-const lm = (options, command) => {
-  command.help()
-}
-
-/**
- * Creates the `netlify lm` command
+ * Creates the `netlify lm:info` command
  * @param {import('../base-command.mjs').default} program
  * @returns
  */
-export const createLmCommand = (program) => {
-  createLmInfoCommand(program)
-  createLmInstallCommand(program)
-  createLmSetupCommand(program)
-  createLmUninstallCommand(program)
-
+export const createLmInfoCommand = (program) =>
   program
-    .command('lm', { hidden: true })
+    .command('lm:info', { hidden: true })
+    .description('Show large media requirements information.')
+    .action(async () => {
+      const { lmInfo } = await import('./lm-info.mjs')
+      await lmInfo()
+    })
+
+/**
+ * Creates the `netlify lm:install` command
+ * @param {import('../base-command.mjs').default} program
+ * @returns
+ */
+export const createLmInstallCommand = (program) =>
+  program
+    .command('lm:install', { hidden: true })
+    .alias('lm:init')
     .description(
-      '[Deprecated and will be removed from future versions] Handle Netlify Large Media operations\nThe lm command will help you manage large media for a site',
+      `Configures your computer to use Netlify Large Media
+It installs the required credentials helper for Git,
+and configures your Git environment with the right credentials.`,
     )
-    .addExamples(['netlify lm:info', 'netlify lm:install', 'netlify lm:setup'])
-    .action(lm)
-}
+    .option('-f, --force', 'Force the credentials helper installation')
+    .action(async ({ force }) => {
+      const { lmInstall } = await import('./lm-install.mjs')
+      await lmInstall({ force })
+    })
+
+/**
+ * Creates the `netlify lm:setup` command
+ * @param {import('../base-command.mjs').default} program
+ * @returns
+ */
+export const createLmSetupCommand = (program) =>
+  program
+    .command('lm:setup', { hidden: true })
+    .description('Configures your site to use Netlify Large Media')
+    .option('-s, --skip-install', 'Skip the credentials helper installation check')
+    .option('-f, --force-install', 'Force the credentials helper installation')
+    .addHelpText('after', 'It runs the install command if you have not installed the dependencies yet.')
+    .action(async (options, command) => {
+      const { lmSetup } = await import('./lm-setup.mjs')
+      await lmSetup(options, command)
+    })
+
+/**
+ * Creates the `netlify lm:uninstall` command
+ * @param {import('../base-command.mjs').default} program
+ * @returns
+ */
+export const createLmUninstallCommand = (program) =>
+  program
+    .command('lm:uninstall', { hidden: true })
+    .alias('lm:remove')
+    .description(
+      'Uninstalls Netlify git credentials helper and cleans up any related configuration changes made by the install command.',
+    )
+    .action(async () => {
+      const { lmUninstall } = await import('./lm-uninstall.mjs')
+      await lmUninstall()
+    })
