+++ src/commands/sites/index.mjs
@@ -1,2 +1,34 @@
-export { sitesCreate } from './sites-create.mjs'
-export { createSitesCommand } from './sites.mjs'
+import {
+  createSitesCreateCommand,
+  createSitesFromTemplateCommand,
+  createSitesListCommand,
+  createSitesDeleteCommand,
+} from './sites.mjs'
+
+/**
+ * The sites command
+ * @param {import('commander').OptionValues} options
+ * @param {import('../base-command.mjs').default} command
+ */
+const sites = (options, command) => {
+  command.help()
+}
+
+/**
+ * Creates the `netlify sites` command
+ * @param {import('../base-command.mjs').default} program
+ * @returns
+ */
+
+export const createSitesCommand = (program) => {
+  createSitesCreateCommand(program)
+  createSitesFromTemplateCommand(program)
+  createSitesListCommand(program)
+  createSitesDeleteCommand(program)
+
+  return program
+    .command('sites')
+    .description(`Handle various site operations\nThe sites command will help you manage all your sites`)
+    .addExamples(['netlify sites:create --name my-new-site', 'netlify sites:list'])
+    .action(sites)
+}
