+++ src/commands/sites/sites-create.mjs
@@ -1,5 +1,4 @@
 // @ts-check
-import { InvalidArgumentError } from 'commander'
 import inquirer from 'inquirer'
 import pick from 'lodash/pick.js'
 import prettyjson from 'prettyjson'
@@ -8,7 +7,7 @@ import { chalk, error, log, logJson, warn } from '../../utils/command-helpers.mj
 import getRepoData from '../../utils/get-repo-data.mjs'
 import { configureRepo } from '../../utils/init/config.mjs'
 import { track } from '../../utils/telemetry/index.mjs'
-import { link } from '../link/index.mjs'
+import { link } from '../link/link.mjs'
 
 export const getSiteNameInput = async (name) => {
   if (!name) {
@@ -143,36 +142,3 @@ export const sitesCreate = async (options, command) => {
 
   return site
 }
-
-const MAX_SITE_NAME_LENGTH = 63
-const validateName = function (value) {
-  // netlify sites:create --name <A string of more than 63 words>
-  if (typeof value === 'string' && value.length > MAX_SITE_NAME_LENGTH) {
-    throw new InvalidArgumentError(`--name should be less than 64 characters, input length: ${value.length}`)
-  }
-
-  return value
-}
-
-/**
- * Creates the `netlify sites:create` command
- * @param {import('../base-command.mjs').default} program
- * @returns
- */
-export const createSitesCreateCommand = (program) =>
-  program
-    .command('sites:create')
-    .description(
-      `Create an empty site (advanced)
-Create a blank site that isn't associated with any git remote. Will link the site to the current working directory.`,
-    )
-    .option('-n, --name <name>', 'name of site', validateName)
-    .option('-a, --account-slug <slug>', 'account slug to create the site under')
-    .option('-c, --with-ci', 'initialize CI hooks during site creation')
-    .option('-m, --manual', 'force manual CI setup.  Used --with-ci flag')
-    .option('--disable-linking', 'create the site without linking it to current directory')
-    .addHelpText(
-      'after',
-      `Create a blank site that isn't associated with any git remote. Will link the site to the current working directory.`,
-    )
-    .action(sitesCreate)
